[Feature]게시글, 댓글 구현

+++ src/main/java/bamboo/entity/PostLikeEntity.java
@@ -21,5 +21,5 @@ public class PostLikeEntity {
     private int userNo;
 
     @Column(name = "post_no")
-    private int postNo;
+    private Long postNo;
 }
